+++ src/pytest_rabbitmq/factories/executor.py
@@ -8,7 +8,10 @@ from mirakuru import TCPExecutor
 class RabbitMqExecutor(TCPExecutor):
     """RabbitMQ executor to start specific rabbitmq instances."""
 
-    def __init__(self, command, host, port, rabbit_ctl, **kwargs):
+    def __init__(
+            self, command, host, port, rabbit_ctl, logpath, path,
+            node_name=None, **kwargs
+    ):  # pylint:disable=too-many-arguments
         """
         Initialize RabbitMQ executor.
 
@@ -16,11 +19,22 @@ class RabbitMqExecutor(TCPExecutor):
         :param str host: host where rabbitmq will be accessible
         :param int port: port under which rabbitmq runs
         :param str rabbit_ctl: rabbitctl location
-        :param dict environ: rabbitmq configuration environment variables
+        :param str logpath:
+        :param str path: Path containing rabbitmq'a mnesia na plugins
+        :param str node_name: RabbitMQ node name
         :param kwargs: see TCPExecutor for description
         """
+        envvars = {
+            'RABBITMQ_LOG_BASE': logpath + f'/rabbit-server.{port}.log',
+            'RABBITMQ_MNESIA_BASE': path + 'mnesia',
+            'RABBITMQ_ENABLED_PLUGINS_FILE': path + 'plugins',
+            'RABBITMQ_NODE_PORT': str(port),
+            # Use the port number in node name, so multiple instances started
+            # at different ports will work separately instead of clustering.
+            'RABBITMQ_NODENAME': node_name or f'rabbitmq-test-{port}',
+        }
         super(RabbitMqExecutor, self).__init__(
-            command, host, port, timeout=60, **kwargs
+            command, host, port, timeout=60, envvars=envvars, **kwargs
         )
         self.rabbit_ctl = rabbit_ctl